Cleaned up CSS and made the LayoutEditor a component
