+++ Frontend/src/style/SeatingStyle.js
@@ -6,8 +6,14 @@ const useStyles = makeStyles((theme) => ({
         width : '100%',
         left: 0,
         right: 0,
-        backgroundColor: "darkgrey",
         
+    },
+    editorMain : {
+        position : 'absolute',
+        width : '100%',
+        left: 0,
+        right: 0,
+        backgroundColor: "darkgrey",
     },
     '@global': {
         '.MuiFormControl-root': {
